+++ styles/pop-overs.styl
@@ -114,18 +114,17 @@ details.popover-container
 
   .emoji
     display: inline-block
-  .button,button
+  .button, button
     display: inline-block
     margin-bottom: 12px
     margin-left: 0
     max-width: initial
+    &:last-child
+      margin: 0
   .button-small
     margin-bottom: 10px
   .button-link
     color: #000
-  .button-link:last-child
-    .button
-      margin: 0
   button:last-child
     margin-bottom: 0
   &.last-section
